Added start of framwork

+++ inc/Particle.hpp
@@ -2,12 +2,12 @@
 #define PARTICLE_HPP
 
 struct Particle {
-  Particle(int in_pos_x, int in_pos_y, double in_radius, double in_mass) :
+  Particle(double in_pos_x, double in_pos_y, double in_radius, double in_mass) :
     pos_x{in_pos_x}, pos_y{in_pos_y}, vel_x{0}, vel_y{0}, radius{in_radius}, mass{in_mass} {}
-  int pos_x;
-  int pos_y;
-  int vel_x;
-  int vel_y;
+  double pos_x;
+  double pos_y;
+  double vel_x;
+  double vel_y;
   double radius;
   double mass;
 };
